feat: improve addon buttons
Adjust styling to increase font size and weight. set a larger consistent height and use flex-grow to make the buttons fill the available space. Also removed padding as with the available space, it shouldn't be necessary

adjust the button text to make it as small as possible. Removing 'addon' as it isn't needed due to context. Renamed Install to web to Install (Web)